Some good changes in Job module (#10)
* Refactor Job module

- Job.start_job/2 now receive a function to execute jobs, so we can write tests easily
without make the full processing of a job
- Turning private methods into public so we can test the complex logic of processing
a job to ensure everything is working

Signed-off-by: Tallys Martins <tallysmartins@gmail.com>

* Remove run_times from measurementd results

Signed-off-by: Tallys Martins <tallysmartins@gmail.com>

* Improvements in docker-compose.yml

- Now, when a docker dependencie is specified, it can define a *wait-for*
attribute in the config.yml file. This attribute will be used to listen the
service port for TCP conections, assuming that its ready if it responds in a
given timeout. There is a script wait-for.sh in the runner-container to make
this check. For example, if a dependencie is given
with wait-for: 5432, the command for the runner service in the docker-compose
is overriden with 'wait-for localhost:5432 -- mix run bench/bench_helper.exs`.
Otherwise, if no dependence is given or the param wait-for is not given, the
default command 'mix run bench/bench_helper' is returned.

Signed-off-by: Tallys Martins <tallysmartins@gmail.com>

* Apply review changes

Signed-off-by: Tallys Martins <tallysmartins@gmail.com>

* Apply review changes

Signed-off-by: Tallys Martins <tallysmartins@gmail.com>

* Apply review changes

Signed-off-by: Tallys Martins <tallysmartins@gmail.com>